Test: 수익률 계산 로직 이동에 따른 테스트 케이스 위치 이동 (LottoPrizeTest > LottoStatisticsTest)
